feat: switch for handleEthTransaction
