+++ packages/shared/src/lib/auxiliary/wallet-connect/utils/index.ts
@@ -1,5 +1,6 @@
 export * from './approveSession'
 export * from './disconnectAllDapps'
 export * from './getAllEvmAddresses'
+export * from './switchToRequiredAccount'
 export * from './updateActiveSessionsToActiveProfile'
 export * from './validateConnectionCodeUri'
